userOpenStore-v1

+++ src/main/java/com/mizore/easybuy/model/dto/UserDTO.java
@@ -12,5 +12,5 @@ public class UserDTO {
 
     private String username;
 
-    private byte role;
+    private int role;
 }
